requirements.txt
openai
sqlalchemy
pandas
python-docx
jira
python-dotenv
crewai
